refactor: changes for balanced artifacts and artifact scaling applied over merge with master

+++ Content.Shared/Xenoarchaeology/Artifact/Components/XenoArtifactNodeComponent.cs
@@ -35,7 +35,7 @@ public sealed partial class XenoArtifactNodeComponent : Component
     public EntityUid? Attached;
 
     [DataField, AutoNetworkedField]
-    public int Budget;
+    public float Budget;
     
     #region Durability
     /// <summary>
